+++ client/src/components/Container/ContainerButton.vue
@@ -1,9 +1,10 @@
 <template>
   <b-list-group-item
     :variant="variant"
-    button
+    :button="!href"
     class="text-center"
-    @click="onClick"
+    :href="href"
+    @click="$emit('click')"
   >
     <slot>
       <i v-if="icon" :class="icon + ' mr-2'" />
@@ -22,14 +23,6 @@ export default {
     textKey: { type: String, default: '' },
     icon: { type: String, default: '' },
     href: { type: String, default: null },
-  },
-  methods: {
-    onClick () {
-      this.$emit('click')
-      if (this.href) {
-        location.href = this.href
-      }
-    },
   },
 }
 </script>