+++ client/src/mixins/ConfirmationDialogue.js
@@ -1,15 +1,17 @@
+import i18n from '@/helper/i18n'
+
 export default {
   methods: {
     confirmationDialogue (messageKey, options = {}) {
       options = Object.assign({
-        title: this.$i18n('are_you_sure'),
+        title: i18n('are_you_sure'),
         okVariant: 'danger',
-        okTitle: this.$i18n('button.delete'),
-        cancelTitle: this.$i18n('button.cancel'),
+        okTitle: i18n('button.delete'),
+        cancelTitle: i18n('button.cancel'),
         centered: true,
         params: {},
       }, options)
-      return this.$bvModal.msgBoxConfirm(this.$i18n(messageKey, options.params), options)
+      return this.$bvModal.msgBoxConfirm(i18n(messageKey, options.params), options)
     },
   },
 }