+++ src/Modules/WallPost/WallPostTransactions.php
@@ -141,7 +141,7 @@ class WallPostTransactions
                     ['href' => '/fairteiler/' . $targetId],
                     [
                         'name' => $this->foodSharePointGateway->getFoodSharePoint($targetId)['name'],
-                        'teaser' => substr($wallPost->body, 0, 100)
+                        'teaser' => substr($wallPost->body ?? '', 0, 100)
                     ],
                     BellType::createIdentifier(BellType::FOOD_SHARE_POINT_POST, $targetId)
                 );