+++ src/RestApi/Models/FoodSharePoint/FoodSharePointPermission.php
@@ -1,0 +1,10 @@
+<?php
+
+namespace Foodsharing\RestApi\Models\FoodSharePoint;
+
+class FoodSharePointPermission
+{
+    public bool $isFollower;
+    public bool $mayEdit;
+    public bool $mayDelete;
+}