+++ tests/Support/Helper/FoodsharingUrl.php
@@ -64,7 +64,7 @@ class FoodsharingUrl extends Db
 
     public function foodSharePointRegionListUrl($region_id): string
     {
-        return '/fairteiler?bid=' . (int)$region_id;
+        return '/region?sub=fairteiler&bid=' . (int)$region_id;
     }
 
     public function foodSharePointGetUrlShort($food_share_point_id): string
@@ -74,12 +74,12 @@ class FoodsharingUrl extends Db
 
     public function foodSharePointGetUrl($food_share_point_id): string
     {
-        return '/fairteiler?sub=ft&id=' . (int)$food_share_point_id;
+        return '/fairteiler/' . (int)$food_share_point_id;
     }
 
     public function foodSharePointEditUrl($food_share_point_id): string
     {
-        return '/fairteiler?sub=edit&id=' . (int)$food_share_point_id;
+        return '/fairteiler/' . (int)$food_share_point_id . '/edit';
     }
 
     public function foodBasketInfoUrl($basket_id): string